requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
Django==3.2.8
django-colorfield==0.7.2
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
fontawesome-free==5.15.4
fontawesomefree==5.15.4
Pillow==9.0.1
pycparser==2.21
pytz==2022.6
sqlparse==0.4.3
